save payment changes

+++ immunitet/database/migrations/2022_09_23_102450_create_company_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
         Schema::create('company', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name', 32);
-            $table->integer('inn');
+            $table->bigInteger('inn');
             $table->integer('postcode');
             $table->foreignId('country_id')
                 ->references('id')
